# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the admin API for the Urchin app.",
        "title": "GoCMS Admin API",
        "contact": {
            "name": "Ricardo",
            "email": "ricardobenthem@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/card": {
            "put": {
                "description": "Updates an existing card with new data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Update an existing card",
                "parameters": [
                    {
                        "description": "Card data to update",
                        "name": "card",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin_app.ChangeCardRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin_app.CardIdResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or could not change card",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/cards": {
            "post": {
                "description": "Adds a new post to the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Add a new post",
                "parameters": [
                    {
                        "description": "Post to add",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin_app.AddCardRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin_app.CardIdResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or missing data",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cards/{schema}": {
            "get": {
                "description": "Retrieves a paginated list of cards by schema UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Get a card list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "schema UUID",
                        "name": "schema",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin_app.GetCardRequest"
                        }
                    },
                    "400": {
                        "description": "Invalid post ID",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/images": {
            "post": {
                "description": "Uploads an image file, saves it, and creates minified versions.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Upload a new image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "The image file to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "A brief description of the image",
                        "name": "excerpt",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin_app.ImageIdResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input, file type, or size",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Server error while saving file",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/images/{name}": {
            "delete": {
                "description": "Deletes an image file from the server by its filename.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Delete an image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Image filename to delete",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin_app.ImageIdResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid or missing filename",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admin_app.AddCardRequest": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Content data of the card\nin: body",
                    "type": "string"
                },
                "image_location": {
                    "description": "Image location URL\nin: body",
                    "type": "string"
                },
                "schema": {
                    "description": "Schema name\nin: body",
                    "type": "string"
                }
            }
        },
...
        "admin_app.CardIdResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID of the card",
                    "type": "string"
                }
            }
        },
        "admin_app.ChangeCardRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID of the card\nin: body",
                    "type": "string"
                },
                "image_location": {
                    "description": "Image location URL\nin: body",
                    "type": "string"
                },
                "json_data": {
                    "description": "JSON data of the card\nin: body",
                    "type": "string"
                },
                "json_schema": {
                    "description": "JSON schema name\nin: body",
                    "type": "string"
                }
            }
        },
...
        "admin_app.GetCardRequest": {
            "type": "object",
            "required": [
                "schema"
            ],
            "properties": {
                "limit": {
                    "description": "Limit number of cards to return\nin: query",
                    "type": "integer"
                },
                "page": {
                    "description": "Page number for pagination\nin: query",
                    "type": "integer"
                },
                "schema": {
                    "description": "Schema name to filter cards\nin: path\nrequired: true",
                    "type": "string"
                }
            }
        },
...
        "admin_app.ImageIdResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID of the image",
                    "type": "string"
                }
            }
        },
...
        "common.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}